+++ runtimes/runtimes/chat/baseChat.ts
@@ -88,6 +88,10 @@ import {
     ListAvailableModelsParams,
     ListAvailableModelsResult,
     listAvailableModelsRequestType,
+    SubscriptionDetailsParams,
+    subscriptionDetailsNotificationType,
+    subscriptionUpgradeNotificationType,
+    SubscriptionUpgradeParams,
 } from '../../protocol'
 import { Chat } from '../../server-interface'
 
@@ -242,5 +246,13 @@ export class BaseChat implements Chat {
 
     public onListAvailableModels(handler: RequestHandler<ListAvailableModelsParams, ListAvailableModelsResult, void>) {
         this.connection.onRequest(listAvailableModelsRequestType.method, handler)
+    }
+
+    public sendSubscriptionDetails(params: SubscriptionDetailsParams) {
+        this.connection.sendNotification(subscriptionDetailsNotificationType.method, params)
+    }
+
+    public onSubscriptionUpgrade(handler: NotificationHandler<SubscriptionUpgradeParams>) {
+        this.connection.onNotification(subscriptionUpgradeNotificationType.method, handler)
     }
 }