+++ types/chat.ts
@@ -23,6 +23,8 @@ export const TAB_BAR_ACTION_REQUEST_METHOD = 'aws/chat/tabBarAction'
 export const CHAT_OPTIONS_UPDATE_NOTIFICATION_METHOD = 'aws/chat/chatOptionsUpdate'
 export const PROMPT_INPUT_OPTION_CHANGE_METHOD = 'aws/chat/promptInputOptionChange'
 export const OPEN_FILE_DIALOG_METHOD = 'aws/chat/openFileDialog'
+export const EXECUTE_SHELL_COMMAND_SHORTCUT_METHOD = 'aws/chat/executeShellCommandShortCut'
+
 // context
 export const CONTEXT_COMMAND_NOTIFICATION_METHOD = 'aws/chat/sendContextCommands'
 export const CREATE_PROMPT_NOTIFICATION_METHOD = 'aws/chat/createPrompt'
@@ -53,6 +55,11 @@ export const LIST_AVAILABLE_MODELS_REQUEST_METHOD = 'aws/chat/listAvailableModel
 
 // button ids
 export const OPEN_WORKSPACE_INDEX_SETTINGS_BUTTON_ID = 'open-settings-for-ws-index'
+
+// Subscription Tiers
+export const SUBSCRIPTION_DETAILS_NOTIFICATION_METHOD = 'aws/chat/subscription/details'
+export const SUBSCRIPTION_UPGRADE_NOTIFICATION_METHOD = 'aws/chat/subscription/upgrade'
+export const SUBSCRIPTION_SHOW_COMMAND_METHOD = 'aws/chat/subscription/show'
 
 export interface ChatItemAction {
     pillText: string
@@ -176,6 +183,23 @@ export interface ChatMessage {
     codeReference?: ReferenceTrackerInformation[]
     fileList?: FileList
     contextList?: FileList
+    quickSettings?: {
+        type: 'select' | 'checkbox' | 'radio'
+        description?: string
+        descriptionLink?: {
+            id: string
+            text: string
+            destination: string
+        }
+        messageId: string
+        tabId: string
+        options: {
+            id: string
+            label: string
+            value: string
+            selected?: boolean | undefined
+        }[]
+    }
 }
 
 /**
@@ -275,6 +299,16 @@ export interface ChatOptions {
      */
     export?: boolean
 
+    /**
+     * Server signals to Chat Client support of show logs feature.
+     */
+    showLogs?: boolean
+
+    /**
+     * Server signals to Client and Chat Client that it supports subscription tier operations
+     */
+    subscriptionDetails?: boolean
+
     /*
         Server signals to Chat Client support of Chat notifications.
         Currently used for sending chat notifications for developer profile updates.
@@ -367,6 +401,7 @@ export interface ButtonClickParams {
     tabId: string
     messageId: string
     buttonId: string
+    metadata: Record<string, string>
 }
 
 export interface ButtonClickResult {
@@ -651,7 +686,7 @@ export interface McpServerClickResult extends McpServerClickParams {
     }
 }
 
-export type TabBarAction = 'export'
+export type TabBarAction = 'export' | 'show_logs'
 export interface TabBarActionParams {
     tabId?: string
     action: TabBarAction
@@ -693,3 +728,14 @@ export interface ListAvailableModelsResult {
 export interface ExecuteShellCommandParams {
     id: string
 }
+
+export interface SubscriptionDetailsParams {
+    subscriptionTier: string
+    subscriptionPeriodReset: Date
+    isOverageEnabled: boolean
+    queryUsage: number
+    queryLimit: number
+    queryOverage: number
+}
+
+export interface SubscriptionUpgradeParams {}